+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added add/remove/toggle class for Tab
 - ContextMenu.radiomenuitem (https://github.com/edvin/tornadofx/issues/646)
 - mutableList can now be bound to an ObservableMap
+- CssSelectionBlock now has all relation selectors (see `CssSubRule.Relation`)
 
 ## [1.7.15]